+++ app/src/main/java/com/surveasy/surveasy/presentation/main/my/setting/SettingViewModel.kt
@@ -2,9 +2,12 @@ package com.surveasy.surveasy.presentation.main.my.setting
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.kakao.sdk.user.UserApiClient
 import com.surveasy.surveasy.app.DataStoreManager
 import com.surveasy.surveasy.domain.base.BaseState
 import com.surveasy.surveasy.domain.usecase.SignoutUseCase
+import com.surveasy.surveasy.domain.usecase.WithdrawUseCase
+import com.surveasy.surveasy.presentation.util.ErrorMsg
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.channels.BufferOverflow
 import kotlinx.coroutines.flow.MutableSharedFlow
@@ -19,6 +22,7 @@ import javax.inject.Inject
 class SettingViewModel @Inject constructor(
     private val dataStoreManager: DataStoreManager,
     private val signoutUseCase: SignoutUseCase,
+    private val withdrawUseCase: WithdrawUseCase,
 ) : ViewModel() {
     private val _events = MutableSharedFlow<SettingUiEvents>(
         replay = 0,
@@ -42,8 +46,36 @@ class SettingViewModel @Inject constructor(
         }.launchIn(viewModelScope)
     }
 
-    fun navigateToWithdraw() {
-        viewModelScope.launch { _events.emit(SettingUiEvents.Withdraw) }
+    fun withdraw() {
+        withdrawUseCase().onEach {
+            when (it) {
+                is BaseState.Success -> {
+                    when(it.data.authProvider){
+                        EMAIL -> Unit
+                        KAKAO -> kakaoUnlink()
+                        else -> Unit
+                    }
+
+                    dataStoreManager.deleteAccessToken()
+                    dataStoreManager.deleteRefreshToken()
+                    dataStoreManager.deleteAutoLogin()
+                    _events.emit(SettingUiEvents.Withdraw)
+                }
+
+                else -> {
+                    _events.emit(SettingUiEvents.ShowSnackBar(ErrorMsg.WITHDRAW_ERROR))
+                }
+            }
+        }.launchIn(viewModelScope)
+    }
+
+    private fun kakaoUnlink() {
+        UserApiClient.instance.unlink { }
+    }
+
+    companion object {
+        const val EMAIL = "EMAIL"
+        const val KAKAO = "KAKAO"
     }
 }
